EWR group generator gets it's own file.
